backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
numpy==1.24.3
